Updating correct version of xen

+++ Xen_Devel/xen_stable_4.11/xen/include/asm-arm/arm64/bitops.h
@@ -24,13 +24,12 @@ static /*__*/always_inline unsigned long __ffs(unsigned long word)
 
 static inline int flsl(unsigned long x)
 {
-        uint64_t ret;
+        int ret;
 
         if (__builtin_constant_p(x))
                return generic_flsl(x);
 
         asm("clz\t%0, %1" : "=r" (ret) : "r" (x));
-
         return BITS_PER_LONG - ret;
 }
